track_odometry: use double instead of float (#455)

+++ track_odometry/src/tf_projection_node.cpp
@@ -111,7 +111,7 @@ public:
 
     if (project_posture_)
     {
-      const float yaw = tf2::getYaw(trans.getRotation());
+      const double yaw = tf2::getYaw(trans.getRotation());
       trans.setRotation(tf2::Quaternion(tf2::Vector3(0.0, 0.0, 1.0), yaw));
     }
 
@@ -123,7 +123,7 @@ public:
     geometry_msgs::TransformStamped trans_out = tf2::toMsg(result);
     if (flat_)
     {
-      const float yaw = tf2::getYaw(trans_out.transform.rotation);
+      const double yaw = tf2::getYaw(trans_out.transform.rotation);
       trans_out.transform.rotation = tf2::toMsg(tf2::Quaternion(tf2::Vector3(0.0, 0.0, 1.0), yaw));
     }
     trans_out.child_frame_id = projected_frame_;
